Changed Double to Long in amounts, added difficulty to blocks, created functions to convert from java objects to protobuf objects and vice-versa, and fixed bugs

+++ src/main/java/pt/up/fc/dcc/ssd/auctionblockchain/addBlock.java
@@ -4,9 +4,10 @@ import java.util.HashMap;
 
 public class addBlock {
     //class to manage adding blocks and checking transactions by miners
-    HashMap<String, Double> fundsTracking;
+    HashMap<String, Long> fundsTracking;
     Block newBlock;
     Wallet minerWallet;
+
     public addBlock(Wallet minerWallet){
         this.minerWallet = minerWallet;
         newBlock = new Block(BlockChain.getLastHash());
